nano-e-commerce/services/user-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
grpcio==1.59.3
grpcio-tools==1.59.3
grpcio-reflection==1.59.3
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
uvloop==0.19.0
python-multipart==0.0.6
structlog==23.2.0
